refactor movie popup and dropdown to hooks
