+++ src/app/components/navigation/NavBar.tsx
@@ -1,10 +1,8 @@
 "use client";
 
 import * as React from "react";
-// import Link from "next/link"
 
 import { cn } from "@/app/lib/utils";
-// import { Icons } from "@/components/icons"
 import {
   NavigationMenu,
   NavigationMenuContent,
