+++ src/app/pages/Home/Home.tsx
@@ -1,4 +1,3 @@
-// import { useAuthenticator } from "@aws-amplify/ui-react";
 import { SearchBar } from "@/app/components/search/SearchBar";
 import {
   QueryClient,
@@ -8,7 +7,6 @@ import {
 const queryClient = new QueryClient();
 
 export const Home = () => {
-  // const { user } = useAuthenticator((context) => [context.user]);
 
   return (
     <>
