+++ src/app/pages/Results/Results.tsx
@@ -3,14 +3,7 @@ import { ResultsCards } from "@/app/components/movie-cards/search-results/Result
 import { useQuery } from "@tanstack/react-query";
 import { useParams } from "react-router";
 
-// import type { Schema } from '../../../../amplify/data/resource'
-// import { generateClient } from 'aws-amplify/api'
-
-
-// const client = generateClient<Schema>()
-
 export const ResultsPage = () => {
-  // const { user } = useAuthenticator((context) => [context.user]);
   const { searchedMovie } = useParams();
 
   if (!searchedMovie) {
@@ -37,7 +30,6 @@ export const ResultsPage = () => {
       <div>Results</div>
       <div>You searched for... {searchedMovie}</div>
       <div>There were {data?.totalResults} results</div>
-      {/* <button onClick={createMovie}>Add new movie</button> */}
       <ResultsCards data={data} />
     </>
   );
